code spliting on all pages except home page
