+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@floating-ui/react": "^0.26.6",
     "@headlessui/react": "^1.7.19",
+    "@loadable/component": "^5.16.3",
     "@tawk.to/tawk-messenger-react": "^2.0.2",
     "flexsearch": "^0.7.43",
     "he": "^1.2.0",
@@ -51,6 +52,7 @@
     "@testing-library/react": "^14.3.1",
     "@types/he": "^1.2.3",
     "@types/html-to-text": "^9.0.4",
+    "@types/loadable__component": "^5.13.9",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@types/react-highlight-words": "^0.20.0",
